Renamed to LearnChinese, added file name as command line input
